+++ sphinx/source/changelog.rst
@@ -17,6 +17,8 @@ Changelog
 * Added the ``--ini`` and ``--show-config`` options to :ref:`tap2sna.py` (for
   setting the value of a configuration parameter and for showing all
   configuration parameter values)
+* Added the ``--tape-analysis`` option to :ref:`tap2sna.py` (for showing an
+  analysis of the tape's tones, pulse sequences and data blocks)
 * :ref:`snapinfo.py` now shows the value of the T-states counter in SZX and Z80
   snapshots
 * Fixed the bug that prevents :ref:`snapinfo.py` from displaying the value of a